Merge pull request #46 from SantiagoTaylor/santi

+++ UI/Webforms/frmEstablishments.aspx.cs
@@ -14,59 +14,37 @@ namespace UI.Webforms
     {
         protected void Page_Load(object sender, EventArgs e)
         {
-            LoadEstablishments();
+            if (!IsPostBack)
+            {
+                LoadEstablishments();
+            }
         }
 
         private void LoadEstablishments()
         {
-            foreach (DataRow r in BLL_Establishment.GetEstablishments().Rows)
+            RepeaterEstablishments.DataSource = BLL_Establishment.GetEstablishments();
+            RepeaterEstablishments.DataBind();
+        }
+
+        protected void Button_Command(object sender, CommandEventArgs e)
+        {
+            string idEstablishment = e.CommandArgument.ToString();
+
+            if (e.CommandName == "Delete")
             {
-                TableRow tr = new TableRow();
-                tr.CssClass = "tablerow";
-                tr.Cells.Add(new TableCell { Text = r["idEstablishment"].ToString() });
-                tr.Cells.Add(new TableCell { Text = r["establishmentName"].ToString() });
-                tr.Cells.Add(new TableCell { Text = r["direction"].ToString() });
-                tr.Cells.Add(new TableCell { Text = r["phone"].ToString() });
-                tr.Cells.Add(new TableCell { Text = r["email"].ToString() });
-
-                TableCell actionCell = new TableCell();
-                LinkButton editButton = new LinkButton
-                {
-                    Text = "<i class='bi bi-pencil-square'></i>",
-                    CommandArgument = r["idEstablishment"].ToString()
-                };
-                editButton.Command += EditButton_Command;
-                editButton.CommandName = "Edit";
-                LinkButton deleteButton = new LinkButton
-                {
-                    Text = "<i class='bi bi-trash3-fill'></i>",
-                    CommandArgument = r["idEstablishment"].ToString(),
-                    OnClientClick = $"return confirmAction('delete', '{r["establishmentName"]}');"
-                };
-                deleteButton.Command += DeleteButton_Command;
-                editButton.CommandName = "Delete";
-
-                actionCell.Controls.Add(editButton);
-                actionCell.Controls.Add(deleteButton);
-                tr.Cells.Add(actionCell);
-                TableEstablishments.Rows.Add(tr);
+                BLL_Establishment.DeleteEstablishment(e.CommandArgument.ToString());
+                Response.Redirect(Request.RawUrl);
             }
-        }
-
-        private void DeleteButton_Command(object sender, CommandEventArgs e)
-        {
-            BLL_Establishment.DeleteEstablishment(e.CommandArgument.ToString());
-            Response.Redirect(Request.RawUrl);
-        }
-
-        private void EditButton_Command(object sender, CommandEventArgs e)
-        {
-            Response.Redirect($"frmCreateEstablishment.aspx?id={e.CommandArgument.ToString()}");
+            else if (e.CommandName == "Modify")
+            {
+                Response.Redirect($"frmCreateEstablishment.aspx?id={e.CommandArgument.ToString()}");
+            }
         }
 
         protected void ButtonRegisterEstablishment_Click(object sender, EventArgs e)
         {
             Response.Redirect("frmCreateEstablishment.aspx");
         }
+
     }
 }
